Placed custom LowercaseCharField into a newly created fields.py file
Also, slight correction to the attrbitues of the PronounSeries class in models.py.
